+++ server/internal/services/membership_service.go
@@ -56,13 +56,13 @@ func (ms *membershipService) CreateMembership(req *models.CreateMembershipReques
 	if req.Paid {
 		// If the membership has been discounted, use the discounted rate instead
 		if req.Discounted {
-			err = ss.UpdateBudget(semesterId, float64(semester.MembershipDiscountFee))
+			err = ss.UpdateBudget(semesterId, float32(semester.MembershipDiscountFee))
 			if err != nil {
 				tx.Rollback()
 				return nil, err
 			}
 		} else {
-			err = ss.UpdateBudget(semesterId, float64(semester.MembershipFee))
+			err = ss.UpdateBudget(semesterId, float32(semester.MembershipFee))
 			if err != nil {
 				tx.Rollback()
 				return nil, err
@@ -191,13 +191,13 @@ func (ms *membershipService) UpdateMembership(req *models.UpdateMembershipReques
 		if existingMembership.Paid {
 			// Update semester's budget
 			if existingMembership.Discounted {
-				err = ss.UpdateBudget(existingMembership.SemesterID, -float64(semester.MembershipDiscountFee))
+				err = ss.UpdateBudget(existingMembership.SemesterID, -float32(semester.MembershipDiscountFee))
 				if err != nil {
 					tx.Rollback()
 					return nil, err
 				}
 			} else {
-				err = ss.UpdateBudget(existingMembership.SemesterID, -float64(semester.MembershipFee))
+				err = ss.UpdateBudget(existingMembership.SemesterID, -float32(semester.MembershipFee))
 				if err != nil {
 					tx.Rollback()
 					return nil, err
@@ -211,14 +211,14 @@ func (ms *membershipService) UpdateMembership(req *models.UpdateMembershipReques
 			// Next compare the discounted flag
 			if !req.Discounted && existingMembership.Discounted {
 				// Member is marked as discounted and updating them to not discounted
-				err = ss.UpdateBudget(semester.ID, float64(semester.MembershipFee-semester.MembershipDiscountFee))
+				err = ss.UpdateBudget(semester.ID, float32(semester.MembershipFee-semester.MembershipDiscountFee))
 				if err != nil {
 					tx.Rollback()
 					return nil, err
 				}
 			} else if req.Discounted && !existingMembership.Discounted {
 				// Member is not marked as discounted and updating them to discounted
-				err = ss.UpdateBudget(semester.ID, -float64(semester.MembershipFee-semester.MembershipDiscountFee))
+				err = ss.UpdateBudget(semester.ID, -float32(semester.MembershipFee-semester.MembershipDiscountFee))
 				if err != nil {
 					tx.Rollback()
 					return nil, err
@@ -227,13 +227,13 @@ func (ms *membershipService) UpdateMembership(req *models.UpdateMembershipReques
 		} else {
 			// Existing member has not paid, and we are updating them to paid
 			if req.Discounted {
-				err = ss.UpdateBudget(semester.ID, float64(semester.MembershipDiscountFee))
+				err = ss.UpdateBudget(semester.ID, float32(semester.MembershipDiscountFee))
 				if err != nil {
 					tx.Rollback()
 					return nil, err
 				}
 			} else {
-				err = ss.UpdateBudget(semester.ID, float64(semester.MembershipFee))
+				err = ss.UpdateBudget(semester.ID, float32(semester.MembershipFee))
 				if err != nil {
 					tx.Rollback()
 					return nil, err
